% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOS-package.R
\docType{data}
\name{chag}
\alias{chag}
\alias{county.chag}
\alias{edu.chag}
\alias{hu12.chag}
\alias{hu4.chag}
\alias{hu8.chag}
\alias{state.chag}
\title{CHAG Datasets}
\format{Data frames with all or a subset of the following variables. Number of observations is
equal to the number of zones in each spatial scale (see Details below). Wildcard options
include a subset of the LAGOS spatial scales (* = hu12, hu8, hu4, edu, county, state),
years of measurement (** = 1985, 1990, 1995, 2000, 2005, 2010), and summary statistic or
units (*** = min, max, mean, std for atmospheric deposition and hydrology; *** = ha, pct
for surficial geology)
\itemize{
   \item *_zoneid: unique zone identifier
   \item *_baseflowindex_***:
   \item *_dep_no3_**_***:
   \item *_dep_so4_**_***
   \item *_dep_totaln_**_***
   \item *_groundwaterrecharge_***
   \item *_prism_ppt_30yr_normal_800mm2_annual_***
   \item *_prism_tmax_30yr_normal_800mm2_annual_***
   \item *_prism_tmean_30yr_normal_800mm2_annual_***
   \item *_prism_tmin_30yr_normal_800mm2_annual_***
   \item *_runoff_***
   \item *_saturationoverlandflow_***
   \item *_surficialgeology_alluv_***
   \item *_surficialgeology_beach_***
   \item *_surficialgeology_colluv_***
   \item *_surficialgeology_dec_resid_***
   \item *_surficialgeology_eolian_***
   \item *_surficialgeology_gf_out_***
   \item *_surficialgeology_grus_***
   \item *_surficialgeology_ice_***
   \item *_surficialgeology_lac_***
   \item *_surficialgeology_lac_clay_***
   \item *_surficialgeology_marine_***
   \item *_surficialgeologyy_open_water_***
   \item *_surficialgeology_other_***
   \item *_surficialgeology_peat_mrsh_***
   \item *_surficialgeology_saprol_***
   \item *_surficialgeology_solif_***
   \item *_surficialgeology_solut_***
   \item *_surficialgeology_till_clay_***
   \item *_surficialgeology_till_loam_***
   \item *_surficialgeology_till_oth_***
   \item *_surficialgeology_till_sand_***
}}
\description{
Datasets containing information on climate, hydrology, atmospheric deposition, and
groundwater characteristics calculated at various LAGOS spatial extents (see ?scales).
}
\details{
Original data sources and layers from which each metric was derived can be found in
additional file 5 of Soranno et al. (2015).
}
\references{
Soranno et al. (2015) \href{https://gigascience.biomedcentral.com/articles/10.1186/s13742-015-0067-4}{Gigascience 4:28}; \href{http://www.mrlc.gov/nlcd2011.php}{National Land Cover Database},
}
\keyword{datasets}

