% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOS-package.R
\docType{data}
\name{conn}
\alias{buffer500m.conn}
\alias{conn}
\alias{connectivity}
\alias{county.conn}
\alias{edu.conn}
\alias{hu12.conn}
\alias{hu4.conn}
\alias{hu8.conn}
\alias{iws.conn}
\alias{lakes4ha.buffer500m.conn}
\alias{state.conn}
\title{Connectivity Datasets}
\format{Data frames with all or a subset of the following variables. Number of observations is
equal to the number of zones in each spatial extent (see ?extents). Wildcard options
include a subset of the LAGOS spatial extents (* = lakes4ha_buffer500m, iws,
hu12, hu8, hu4, edu, county, state) and additional wildcards described below.
\itemize{
     \item Identifiers and Zone Metadata
     \itemize{
         \item *_zoneid
         \item *_nwiexclusions
         }
     \item Lake Connectivity Metrics. Lake abundance metrics are derived for all lakes,
     isolated lakes ("isolated" - no inflow or outflow streams), headwater lakes ("headwater" - no inflowing streams),
     stream drainage lakes ("drstream" - lakes with inflowing stream(s), with or without outflows,
     and no upstream lakes ≥10 ha), and lake drainage lakes ("drlakeStream" lakes with inflowing
     stream(s) and at least one upstream lake ≥10 ha).
     \itemize{
         \item *_lakes_**_***: all lakes that are within or intersect the zone boundary
         and are in a particular size class (** = lakes4ha, lakes4to10ha, lakes10ha),
         expressed as average lake area (** = agsize_ha) or count (** = count)
         \item *_lakes_**_overlapping_***: all lakes that are clipped to the zone boundary
         and are in a particular size class (** = lakes4ha, lakes4to10ha, lakes10ha), expressed
         as sum of lake area (*** = area_ha) or percent of zone area (*** = area_pct)
         \item *_lakes_**_contributing_area_ha: all lakes that are within or intersect the zone boundary,
         and are in a particular size class (** = lakes4ha, lakes4to10ha, lakes10ha), expressed as
         sum of lake area
         \item *_lakes_**_isolated_***: all isolated lakes that are within or intersect the zone boundary,
         and are in a particular size class (** = lakes4ha, lakes4to10ha, lakes10ha), expressed
         as average size in hectares (*** = avgsize_ha) or count (*** = count
         \item *_lakes_**_isolated_overlapping_***: all isolated lakes that are clipped to the zone boundary
         and are in a particular size class (** = lakes4ha, lakes4to10ha, lakes10ha), expressed
         as sum of lake area (*** = area_ha) or percent of zone area (*** = area_pct)
         \item *_lakes_**_isolated_contributing_area_ha: all isolated lakes that are within or intersect the zone boundary,
         and are in a particular size class (** = lakes4ha, lakes4to10ha, lakes10ha), expressed as
         sum of lake area
          \item *_lakes_**_headwater_***: all headwater lakes that are within or intersect the zone boundary,
         and are in a particular size class (** = lakes4ha, lakes4to10ha, lakes10ha), expressed
         as average size in hectares (*** = avgsize_ha) or count (*** = count
         \item *_lakes_**_headwater_overlapping_***: all headwater lakes that are clipped to the zone boundary
         and are in a particular size class (** = lakes4ha, lakes4to10ha, lakes10ha), expressed
         as sum of lake area (*** = area_ha) or percent of zone area (*** = area_pct)
         \item *_lakes_**_headwater_contributing_area_ha: all headwater lakes that are within or intersect the zone boundary,
         and are in a particular size class (** = lakes4ha, lakes4to10ha, lakes10ha), expressed as
         sum of lake area
         \item *_lakes_**_drstream_***: all drstream lakes that are within or intersect the zone boundary,
         and are in a particular size class (** = lakes4ha, lakes4to10ha, lakes10ha), expressed
         as average size in hectares (*** = avgsize_ha) or count (*** = count
         \item *_lakes_**_drstream_overlapping_***: all drstream lakes that are clipped to the zone boundary
         and are in a particular size class (** = lakes4ha, lakes4to10ha, lakes10ha), expressed
         as sum of lake area (*** = area_ha) or percent of zone area (*** = area_pct)
         \item *_lakes_**_drstream_contributing_area_ha: all drstream lakes that are within or intersect the zone boundary,
         and are in a particular size class (** = lakes4ha, lakes4to10ha, lakes10ha), expressed as
         sum of lake area
         \item *_lakes_**_drlakestream_***: all drlakestream lakes that are within or intersect the zone boundary,
         and are in a particular size class (** = lakes4ha, lakes4to10ha, lakes10ha), expressed
         as average size in hectares (*** = avgsize_ha) or count (*** = count
         \item *_lakes_**_drlakestream_overlapping_***: all drlakestream lakes that are clipped to the zone boundary
         and are in a particular size class (** = lakes4ha, lakes4to10ha, lakes10ha), expressed
         as sum of lake area (*** = area_ha) or percent of zone area (*** = area_pct)
         \item *_lakes_**_drlakestream_contributing_area_ha: all drlakestream lakes that are within or intersect the zone boundary,
         and are in a particular size class (** = lakes4ha, lakes4to10ha, lakes10ha), expressed as
         sum of lake area
         \item *_latewisconsinglaciation_glaciation: zone glaciation status during
         the late Wisconsin glaciation (Glaciated, Not_Glaciated, Partially)
     }
     \item Stream Connectivity Metrics
     \itemize{
         \item *_streamdensity_streams_**: all streams (NHDFlowline minus artifical lines through lakes)
         within the zone expressed as sum of length in meters (** = sum_lengthm) or density
         in meters per hectare (** = mperha)
         \item *_streamdensity_headwaters_*: headwater streams (1st-3rd Strahler order) within the zone
         expressed as sum of length in meters (** = sum_lengthm) or density
         in meters per hectare (** = mperha)
         \item *_streamdensity_midreaches_*: Midreach streams (4th-6th Strahler order) within the zone
         expressed as sum of length in meters (** = sum_lengthm) or density
         in meters per hectare (** = mperha)
         \item *_streamdensity_rivers_*: large river-streams (>6th Strahler order) within the zone
         expressed as sum of length in meters (** = sum_lengthm) or density
         in meters per hectare (** = mperha)
         \item *_canalditchdensity_**: canals within the zone,
         expressed as sum of length in meters (** = sum_lengthm) or density
         in meters per hectare (** = mperha)
         }
     \item Wetland Connectivity Metrics
     \itemize{
         \item *_wl_allwetlands**_***: ** = undissolved, dissolved, *** = avgsize_ha, count
         \item *_wl_allwetlands**_overlapping_***: ** = undissolved, dissolved; *** = area_ha, area_pct
         \item *_wl_allwetlands**_contributing_area_ha: ** = undissolved, dissolved
         \item *_wl_isolatedwetlandsundissolved_***: ** = undissolved, dissolved, *** = avgsize_ha, count
         \item *_wl_isolatedwetlandsundissolved_overlapping_***: ** = undissolved, dissolved; *** = area_ha, area_pct
         \item *_wl_isolatedwetlandsundissolved_contributing_area_ha: ** = undissolved, dissolved
         \item *_wl_singlewetlandsundissolved_***: ** = undissolved, dissolved, *** = avgsize_ha, count
         \item *_wl_singlewetlandsundissolved_overlapping_***: ** = undissolved, dissolved; *** = area_ha, area_pct
         \item *_wl_singlewetlandsundissolved_contributing_area_ha: ** = undissolved, dissolved
         \item *_wl_connectedwetlandsundissolved_***: ** = undissolved, dissolved, *** = avgsize_ha, count
         \item *_wl_connectedwetlandsundissolved_overlapping_***: ** = undissolved, dissolved; *** = area_ha, area_pct
         \item *_wl_connectedwetlandsundissolved_contributing_area_ha: ** = undissolved, dissolved
         \item *_wl_forestedwetlandsundissolved_***: ** = undissolved, dissolved, *** = avgsize_ha, count
         \item *_wl_forestedwetlandsundissolved_overlapping_***: ** = undissolved, dissolved; *** = area_ha, area_pct
         \item *_wl_forestedwetlandsundissolved_contributing_area_ha: ** = undissolved, dissolved
         \item *_wl_scrubshrubwetlandsundissolved_***: ** = undissolved, dissolved, *** = avgsize_ha, count
         \item *_wl_scrubshrubwetlandsundissolved_overlapping_***: ** = undissolved, dissolved; *** = area_ha, area_pct
         \item *_wl_scrubshrubwetlandsundissolved_contributing_area_ha: ** = undissolved, dissolved
         \item *_wl_openwaterwetlandsundissolved_***: ** = undissolved, dissolved, *** = avgsize_ha, count
         \item *_wl_openwaterwetlandsundissolved_overlapping_***: ** = undissolved, dissolved; *** = area_ha, area_pct
         \item *_wl_openwaterwetlandsundissolved_contributing_area_ha: ** = undissolved, dissolved
         \item *_wl_regimefwetlandsundissolved_***: ** = undissolved, dissolved, *** = avgsize_ha, count
         \item *_wl_regimefwetlandsundissolved_overlapping_***: ** = undissolved, dissolved; *** = area_ha, area_pct
         \item *_wl_regimefwetlandsundissolved_contributing_area_ha: ** = undissolved, dissolved
         \item *_wl_regimegwetlandsundissolved_***: ** = undissolved, dissolved, *** = avgsize_ha, count
         \item *_wl_regimegwetlandsundissolved_overlapping_***: ** = undissolved, dissolved; *** = area_ha, area_pct
         \item *_wl_regimegwetlandsundissolved_contributing_area_ha: ** = undissolved, dissolved
         \item *_wl_regimeawetlandsundissolved_***: ** = undissolved, dissolved, *** = avgsize_ha, count
         \item *_wl_regimeawetlandsundissolved_overlapping_***: ** = undissolved, dissolved; *** = area_ha, area_pct
         \item *_wl_regimeawetlandsundissolved_contributing_area_ha: ** = undissolved, dissolved
         \item *_wl_regimecwetlandsundissolved_***: ** = undissolved, dissolved, *** = avgsize_ha, count
         \item *_wl_regimecwetlandsundissolved_overlapping_***: ** = undissolved, dissolved; *** = area_ha, area_pct
         \item *_wl_regimecwetlandsundissolved_contributing_area_ha: ** = undissolved, dissolved
         \item *_wl_regimehwetlandsundissolved_***: ** = undissolved, dissolved, *** = avgsize_ha, count
         \item *_wl_regimehwetlandsundissolved_overlapping_***: ** = undissolved, dissolved; *** = area_ha, area_pct
         \item *_wl_regimehwetlandsundissolved_contributing_area_ha: ** = undissolved, dissolved
    }
    }}
\description{
Datasets containing lake, stream, and wetland abundance and connectivity metrics calculated
at the 500m buffer (lakes4ha.buffer500m.conn), HUC12 (hu12.conn), HUC8 (hu8.conn),
HUC4 (hu4.conn), EDU (edu.conn), county (county.conn) and state (state.conn) extents.
}
\references{
Soranno et al. (2015) \href{https://gigascience.biomedcentral.com/articles/10.1186/s13742-015-0067-4}{Gigascience 4:28}
}
\keyword{datasets}

