% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOS-package.R
\docType{data}
\name{county}
\alias{county}
\title{County Data}
\format{A data frame with 955 observations of 8 variables:
\itemize{
    \item county_state: abbreviation of state corresponding to the county
    \item county_name: text name of the zone
    \item county_ha_in_nwi: area (ha) of polygon within the USFWS National Wetland Inventory (NWI) boundary
    \item county_ha: surface area (ha) of polygon
    \item county_zoneid: unique zone identifier
    \item county_pct_in_nwi: percent of polygon within the USFWS National Wetland Inventory (NWI) boundary
    \item county_lat: latitude of centroid of each polygon in decimal degrees (NAD83)
    \item county_long: longitude of centroid of each polygon in decimal degrees (NAD83)
    }}
\description{
A dataset containing information about county characteristics for LAGOS lakes.
For details on zone delineations, see additional file 7 in Soranno et al. (2015).
}
\references{
Soranno et al. (2015) \href{https://gigascience.biomedcentral.com/articles/10.1186/s13742-015-0067-4}{Gigascience 4:28}
}
\keyword{datasets}

