% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{display_names}
\alias{display_names}
\title{Convert table names for display}
\usage{
display_names(nms_raw)
}
\description{
Convert table names for display
}
\examples{
\dontrun{
lg <- lagosne_load("1.087.1")
display_names(names(lg$iws))
display_names(names(lg$hu4))
}
}
