% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOS-package.R
\docType{data}
\name{epi.nutr}
\alias{epi.nutr}
\title{Epilimnion Data}
\format{A data frame with 276488 observations and 94 variables:
\itemize{
            \item eventida10541
            \item lagoslakeid
            \item programname
            \item programtype
            \item lagosversion
            \item sampledate
            \item chla
            \item colora
            \item colort
            \item dkn
            \item doc
            \item nh4
            \item no2
            \item no2no3
            \item srp
            \item tdn
            \item tdp
            \item tkn
            \item tn
            \item toc
            \item ton
            \item tp
            \item secchi
            \item chla_qual
            \item colora_qual
            \item colort_qual
            \item dkn_qual
            \item doc_qual
            \item nh4_qual
            \item no2_qual
            \item no2no3_qual
            \item srp_qual
            \item tdn_qual
            \item tdp_qual
            \item tkn_qual
            \item tn_qual
            \item toc_qual
            \item ton_qual
            \item tp_qual
            \item secchi_qual
            \item chla_censorcode
            \item colora_censorcode
            \item colort_censorcode
            \item dkn_censorcode
            \item doc_censorcode
            \item nh4_censorcode
            \item no2_censorcode
            \item no2no3_censorcode
            \item secchi_censorcode
            \item srp_censorcode
            \item tdn_censorcode
            \item tdp_censorcode
            \item tkn_censorcode
            \item tn_censorcode
            \item toc_censorcode
            \item ton_censorcode
            \item tp_censorcode
            \item chla_detectionlimit
            \item colora_detectionlimit
            \item colort_detectionlimit
            \item dkn_detectionlimit
            \item doc_detectionlimit
            \item nh4_detectionlimit
            \item no2_detectionlimit
            \item no2no3_detectionlimit
            \item secchi_detectionlimit
            \item srp_detectionlimit
            \item tdn_detectionlimit
            \item tdp_detectionlimit
            \item tkn_detectionlimit
            \item tn_detectionlimit
            \item toc_detectionlimit
            \item ton_detectionlimit
            \item tp_detectionlimit
            \item chla_labmethodname
            \item colora_labmethodname
            \item colort_labmethodname
            \item dkn_labmethodname
            \item doc_labemthodname
            \item nh4_labmethodname
            \item no2_labmethodname
            \item no2no3_labmethodname
            \item srp_labmethodname
            \item tdn_labmethodname
            \item tdp_labmethodname
            \item tkn_labmethodname
            \item tn_labmethodname
            \item toc_labmethodname
            \item tp_labmethodname
            \item chla_methodinfo
            \item nh4_methodinfo
            \item secchi_methodinfo
            \item sampleyear
            \item samplemonth
    }}
\description{
A dataset containing physical and chemical attributes of the lakes' epilimnion (surface waters).
}
\references{
Soranno et al. (2015) \href{https://gigascience.biomedcentral.com/articles/10.1186/s13742-015-0067-4}{Gigascience 4:28}
}
\keyword{datasets}

