% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOS-package.R
\docType{data}
\name{hu12}
\alias{hu12}
\title{HUC 12 Data}
\format{A data frame with 20257 observations of 12 variables:
\itemize{
    \item hu12:
    \item hu12_name:
    \item hu12_country:
    \item hu12_pct_in_us:
    \item hu12_zoneid:
    \item hu12_pct_in_nwi:
    \item hu12_ha:
    \item hu12_ha_in_nwi:
    \item hu12_pct_in_usa:
    \item hu12_ha_in_usa:
    \item hu12_lat:
    \item hu12_long:
    }}
\description{
A dataset containing information about the HUC 12 characteristics for LAGOS lakes.
For details on hydrologic units, refer to the USGS hydrologic unit code documentation.
}
\references{
\href{http://water.usgs.gov/GIS/huc.html}{USGS Hydrologic Unit Maps}
}
\keyword{datasets}

