% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{identDector}
\alias{identDector}
\alias{identDetector}
\title{Not sure what this does}
\usage{
identDetector(list)
}
\arguments{
\item{list}{vector of data.frame names}
}
\value{
vector
}
\description{
given a vector contains dataframes names return a vector of dataframes identifiers
}

