% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{info_table}
\alias{info_table}
\title{Summarize all LAGOS flat files}
\usage{
info_table(lg)
}
\arguments{
\item{lg}{list output of lagos_load}
}
\description{
Generate table summary statistics for all LAGOS data frames.
}
\examples{
\dontrun{
lg <- lagos_load("1.087.1")
LAGOS:::info_table(lg)
}
}
\author{
Masrour Farzan
}
