% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOS-package.R
\docType{data}
\name{iws.lulc}
\alias{iws.lulc}
\title{Watershed LULC Data}
\format{A data frame with 51065 observations of 158 variables:
\itemize{
   \item iws_nhdid
   \item iws_canopy2001_*: per pixel tree canopy percent for zone, 2001. Summary statistics include minimum, maximum, mean and standard deviation (* = min, max, mean, std).
   \item iws_damdensity_*: dams in zone. Summary statistics include a point count, points per hectare, and points per square kilometer (* = pointcount, poinsperha, pointspersqkm)
   \item iws_nlcd*_**_0: No data
   \item iws_nlcd*_**_11: Open water
   \item iws_nlcd*_**_21: Residential, low intensity
   \item iws_nlcd*_**_22: Residential, high intensity
   \item iws_nlcd*_**_23: Commercial/industrial/transportation
   \item iws_nlcd*_**_24: Developed, high Intensity. Note, this variable was not calculated in 1992 (* = 2001, 2006, 2011)
   \item iws_nlcd*_**_31: Bare rock/sand/clay
   \item iws_nlcd1992_**_32: Quarries/strip mines/gravel pits. Note, this variable only calculated in 1992.
   \item iws_nlcd1992_**_33: Transitional (barren). Note, this variable only calculated in 1992.
   \item iws_nlcd*_**_41: Deciduous forest
   \item iws_nlcd*_**_42: Evergreen forest
   \item iws_nlcd*_**_43: Mixed forest
   \item iws_nlcd1992_**_51: Shrubland. Note, this variable only calculated in 1992.
   \item iws_nlcd1992_**_52: Scrub/shrub. Note, this variable not calculated in 1992 (* = 2001, 2006, 2011)
   \item iws_nlcd1992_**_61: Orchards/vineyards/other. Note, this variable only calculated in 1992.
   \item iws_nlcd*_**_71: Grasslands/herbaceous
   \item iws_nlcd*_**_81: Pasture/hay
   \item iws_nlcd*_**_82: Row crops
   \item iws_nlcd1992_**_83: Small grains. Note, this variable only calculated in 1992.
   \item iws_nlcd1992_**_85: Urban/recreational grasses. Note, this variable only calculated in 1992.
   \item iws_nlcd*_**_90: Woody wetlands. Note, this variable not calculated in 1992 (* = 2001, 2006, 2011). Equivalent code for 1992 = 91.
   \item iws_nlcd1992_**_91: Woody wetlands. Note, this variable only calculated in 1992. Equivalent code for 2001, 2006 and 2011 = 90.
   \item iws_nlcd1992_**_92: Emergent herbaceous wetlands. Note, this variable only calculated in 1992. Equivalent code for 2001, 2006, and 2011 = 95.
   \item iws_nlcd*_**_95: Emergent herbaceous wetlands. Note, this variable not calculated in 1992. Equivalent code for 1992 = 92.
   \item iws_roaddensity_*: Roads within the zone, calculated as sum of the length in meters, or sum of the length in meters divided by area of zone in hectares (* = sum_lengthm, density_mperha)
   \item iws_slope_*: The slope at each cell is the slope with respect to its immediate neighbors (10m scale). Cells in IWS summarized as either the minimum, maximum, mean, or standard deviation (* = min, max, mean, std)
   \item iws_tri_*: Terrain ruggedness index (TRI) at each cell in meters with repsect to its immediate neighbors (10m scale). Cells in IWS summarized as either the minimum, maximum, mean, or standard deviation (* = min, max, mean, std)
   \item iws_zoneid: unique zone identifier
   \item iws_lagoslakeid: unique integer identifier for each lake in LAGOS
   \item lakeconnection:
   \item borderiws:
    }}
\description{
A dataset containing information on the land use and land cover (LULC) characteristics calculated at the interlake watershed spatial extent.
For details on how the watersheds were delineated, see additional file 8 in Soranno et al. (2015).
Original data sources and layers from which each metric was derived can be found in additional file 5 of Soranno et al. (2015).
}
\details{
Note the variable descriptions below for each National Land Cover Database (NLCD)
code are not repeated for each year and statistic combination.
NLCD data come from multiple years (* = 1992, 2001, 2006 and 2011) and were calculated as either the total area in hectares of coverage or percent coverage (** = ha, pct).
}
\references{
Soranno et al. (2015) \href{http://www.mrlc.gov/nlcd2011.php}{National Land Cover Database},
}
\keyword{datasets}

