% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOS-package.R
\docType{data}
\name{iws}
\alias{iws}
\title{Watershed Data}
\description{
A dataset containing information on watershed characteristics for LAGOS lakes
}
\details{
\itemize{
    \item iws_nhdid
    \item iws_areaha
    \item iws_perimkm
    \item iws_lat latitude
    \item iws_long longitude
    \item iws_lakeareaha
    \item iws_pct_in_usa
    \item iws_ha_in_usa
    \item iws_pct_in_nwi
    \item iws_ha_in_nwi
    \item iws_zoneid
    \item iws_lagoslakeid
    }
}
\keyword{datasets}

