% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOS-package.R
\docType{package}
\name{LAGOS-package}
\alias{LAGOS-package}
\alias{LAGOS}
\title{R interface to the LAGOS lakes database}
\description{
R interface to the LAGOS lakes database
}
\author{
\email{stachel2@msu.edu}
}
