% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{lagos_get}
\alias{lagos_get}
\title{lagos_get}
\usage{
lagos_get(dest_folder = NA)
}
\arguments{
\item{dest_folder}{file.path optional will default to the location returned by \code{\link[rappdirs]{user_data_dir}}.}
}
\description{
Retrieves LAGOS flat files from EDI.
}
\examples{
\dontrun{
lagos_get()
}
}
