% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{lagos_get_collins_2017}
\alias{lagos_get_collins_2017}
\title{Get LAGOSNE snapshot from Collins et al. 2017}
\usage{
lagos_get_collins_2017(dest_folder, overwrite = FALSE)
}
\arguments{
\item{dest_folder}{file.path not implemented yet}

\item{overwrite}{logical overwrite existing data?}
}
\description{
Get LAGOSNE snapshot from Collins et al. 2017
}
\examples{
\dontrun{
lagos_get_collins_2017()
}
}
\references{
Collins S., S. Oliver, J. Lapierre, E. Stanley, J. Jones,
 T. Wagner, P. Soranno. 2016. LAGOS - Lake nitrogen, phosphorus,
 stoichiometry, and geospatial data for a 17-state region of the U.S..
 Environmental Data Initiative.
 doi:10.6073/pasta/3abb4a56e76a52a12a366a338fc07dd8.
}
