% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{lagos_get_oliver_2015}
\alias{lagos_get_oliver_2015}
\title{Get depth data}
\usage{
lagos_get_oliver_2015(dest_folder)
}
\arguments{
\item{dest_folder}{file.path not implemented yet}
}
\description{
Get depth data from Oliver et al. 2015
}
\examples{
\dontrun{
lagos_get_oliver_2015()
}
}
\references{
Oliver SK, Soranno PA, Fergus EC, Wagner T, Webster KE, Scott C,
 Winslow LA, Downing JA, Stanley EH. 2015. LAGOS - Predicted and observed
 maximum depth values for lakes in a 17-state region of the U.S. Long Term
 Ecological Research Network.
 doi:10.6073/pasta/f00a245fd9461529b8cd9d992d7e3a2f
}
