% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{lagos_ingest}
\alias{lagos_ingest}
\title{Ingest LAGOS flat files}
\usage{
lagos_ingest(version)
}
\arguments{
\item{version}{character LAGOS database version string}
}
\description{
Ingest LAGOS data from component flat files
}
\examples{
\dontrun{
lagos_ingest("1.054.1")
}

}

