% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOS1.054.1.R
\name{lagos_load}
\alias{lagos_load}
\title{lagos_load}
\usage{
lagos_load(version)
}
\arguments{
\item{version}{character LAGOS database version string}
}
\description{
lagos_load
}
\examples{
\dontrun{
lagos_load("1.054.1")
}
}

