% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{lagos_load_collins_2017}
\alias{lagos_load_collins_2017}
\title{Load LAGOSNE snapshot from Collins et al. 2017.}
\usage{
lagos_load_collins_2017(fpath = NA)
}
\arguments{
\item{fpath}{file.path optionally specify custom location of csv data file}
}
\description{
Load LAGOSNE snapshot from Collins et al. 2017.
}
\examples{
\dontrun{
lg <- lagos_load_collins_2017()
}
}
\references{
Collins S., S. Oliver, J. Lapierre, E. Stanley, J. Jones,
 T. Wagner, P. Soranno. 2016. LAGOS - Lake nitrogen, phosphorus,
 stoichiometry, and geospatial data for a 17-state region of the U.S..
 Environmental Data Initiative.
 doi:10.6073/pasta/3abb4a56e76a52a12a366a338fc07dd8.
}
