% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_group.R
\name{lagos_select_group}
\alias{lagos_select_group}
\title{lagos_select_group}
\usage{
lagos_select_group(dt, table_column_nested)
}
\arguments{
\item{dt}{data.frame}

\item{table_group_nested}{nested list of tables and groups of columns to select from LAGOS. See details for
a list of possible table and group combinations}
}
\description{
Select and merge columns based on a nested list of LAGOS tables and groups of variables.
}
\details{
This function takes advantage of predefined groups of variables within each LAGOS table.
The possible table and group combinations include:
\enumerate{
Groups for the "limno" tables include water.quality (all chemistry + chlorophyll + water clarity measures).
\item Groups for "geo" .chag tables include deposition (all atmospheric deposition related variables),
hydrology (basefloe, groundwater, runoff and overland flow variables), climate (30-year normal precipitation and temperature variables),
and geology (all surficial geology variables)
\item Groups for the "geo" .lulc tables include topography (Terrain Roughness Index and slope variables),
lulc1992 (all LULC variables from 1992), lulc2001, lulc2006, lulc2011.
\item Groups for the "geo" .conn tables include lakes (all lake connectivity metrics),
wetlands (all wetland connectivity metrics) and streams (all stream/river connectivity metrics)
}
}
\examples{
\dontrun{
dt <- lagos_load("1.054.1")

table_columns <- list("epi.nutr" = c("waterquality"), "iws.lulc" = c("iws_nlcd2011_pct_95"))

dt_reduced <- LAGOS::lagos_select(dt, table_columns)
}
}

