% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSNE-package.R
\docType{data}
\name{lagos_source_program}
\alias{lagos_source_program}
\alias{programs}
\alias{lagos.source.programs}
\alias{source.programs}
\alias{lagos.source}
\title{LAGOSNE sources}
\format{
A data frame with observations of 19 variables:
\itemize{
     \item programid: unique LAGOSNE program identifier
     \item programname: name of the sampling/monitoring program that
     collected the data
     \item metadataid: unique metadata identifier associated with
     \item sourceid
     \item sourcename
     \item composite
     \item programtype: categorical description of the type of sampling
     /monitoring effort
     \item fundingsource
     \item labtype
     \item programlink
     \item programstatus
     \item dataprovider
     \item organization
     \item sourcedescription
     \item beg_yr
     \item end_yr
     \item state_name
     \item lakecount
     \item variable_list
     }
}
\description{
A dataset identifying the sources of limnological data in LAGOSNE.
}
\references{
Soranno et al. (2017) \href{https://doi.org/10.1093/gigascience/gix101}{Gigascience 6:12}
}
\keyword{datasets}
