% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{LAGOSNE-deprecated}
\alias{LAGOSNE-deprecated}
\alias{lagos_get}
\alias{lagos_load}
\alias{lagos_select}
\alias{lagos_compile}
\alias{LAGOSNE-deprecated-package}
\title{Deprecated functions in the LAGOSNE package}
\usage{
lagos_get(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older versions of
the LAGOSNE package. They may eventually be completely removed.
}
\section{Details}{

\tabular{rl}{
  \code{lagos_get} \tab now a synonym for \code{\link{lagosne_get}}\cr
  \code{lagos_load} \tab now a synonym for \code{\link{lagosne_load}}\cr
  \code{lagos_select} \tab now a synonym for \code{\link{lagosne_select}}\cr
  \code{lagos_compile} \tab now a synonym for \code{\link{lagosne_compile}}\cr
}
}

