% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSNE-package.R
\docType{package}
\name{LAGOSNE-package}
\alias{LAGOSNE-package}
\alias{LAGOSNE}
\title{R interface to the LAGOSNE lakes database}
\description{
Interface to the LAGOSNE lakes database
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cont-limno/LAGOSNE}
  \item \url{https://cont-limno.github.io/LAGOSNE/}
  \item Report bugs at \url{https://github.com/cont-limno/LAGOSNE/issues}
}

}
\author{
\email{stachel2@msu.edu}
}
\keyword{internal}
