% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{lagosne_get}
\alias{lagosne_get}
\title{Retrieve LAGOSNE flat files}
\usage{
lagosne_get(version, overwrite = FALSE, dest_folder = NA)
}
\arguments{
\item{version}{character LAGOSNE database version string}

\item{overwrite}{logical overwrite existing data for the specified version}

\item{dest_folder}{file.path not implemented yet
by \code{\link[rappdirs]{user_data_dir}}.}
}
\description{
Retrieve LAGOSNE flat files from EDI.
}
\examples{
\dontrun{
lagosne_get(version = "1.087.1")
}
}
