% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOS-package.R
\docType{data}
\name{lakes.limno}
\alias{lakes.limno}
\alias{meandepth}
\alias{maxdepth}
\alias{legacyid}
\title{Water Quality Lakes}
\format{A data frame with 141265 observations of 8 variables:
\itemize{
            \item lagoslakeid:	    unique lake identifier developed for LAGOS-NE
            \item nhdid:	          the 'Permanent_identifier' from the NHD, but is called the NHDID in LAGOS-NE
            \item lagosname1:	    name of lake from source dataset used as the
            authoritative lake name for LAGOS
            \item meandepth:	      mean lake depth (meters)
            \item meandepthsource:	source of mean depth data, see "lagos.source.program" for definitions
            \item maxdepth:	      maximum lake depth (meters)
            \item maxdepthsource:	source of maximum depth data, see "lagos.source.program" for definitions
            \item legacyid:	      unique lake identifier from the original water quality source dataset, this is not standardized and each source dataset has its own system
    }}
\description{
A dataset containing depth and identifying information on lakes in the limno module.
}
\references{
Soranno et al. (2015) \href{https://gigascience.biomedcentral.com/articles/10.1186/s13742-015-0067-4}{Gigascience 4:28}
}
\keyword{datasets}
