% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSNE-package.R
\docType{data}
\name{lakes_limno}
\alias{lakes_limno}
\alias{meandepth}
\alias{maxdepth}
\alias{legacyid}
\alias{depth}
\title{Metadata for Lakes with Water Quality}
\format{
A data frame with observations of 8 variables:
\itemize{
            \item lagoslakeid:	    unique lake identifier developed for
            LAGOS-NE
            \item nhdid:	          the 'Permanent_identifier' from the NHD,
            but is called the NHDID in LAGOS-NE
            \item lagosname1:	    name of lake from source dataset used as
            the authoritative lake name for LAGOSNE
            \item meandepth:	      mean lake depth (meters)
            \item meandepthsource:	source of mean depth data, see
            "lagos_source_program" for definitions
            \item maxdepth:	      maximum lake depth (meters)
            \item maxdepthsource:	source of maximum depth data, see
            "lagos_source_program" for definitions
            \item legacyid:	      unique lake identifier from the original
            water quality source dataset, this is not standardized and each
            source dataset has its own system
    }
}
\description{
A dataset containing depth and names for lakes included in the limno
module.
}
\references{
Soranno et al. (2017) \doi{10.1093/gigascience/gix101}
}
\seealso{
\code{\link[LAGOSNE]{lakes.geo}}
}
\keyword{datasets}
