% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSNE-package.R
\docType{data}
\name{lg_extent}
\alias{lg_extent}
\title{LAGOSNE extent}
\format{An object of class \code{sfc_POLYGON} (inherits from \code{sfc}) of length 1.}
\usage{
lg_extent
}
\description{
A simple features polygons representing the geographical extent of the LAGOSNE dataset
}
\references{
Soranno et al. (2015) \href{https://gigascience.biomedcentral.com/articles/10.1186/s13742-015-0067-4}{Gigascience 4:28}
}
\keyword{datasets}
