% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSNE-package.R
\docType{data}
\name{lg_subset}
\alias{lg_subset}
\title{LAGOSNE subset}
\format{An object of class \code{list} of length 26.}
\usage{
lg_subset
}
\description{
A dataset with the first two rows of many of the LAGOSNE tables
}
\references{
Soranno et al. (2017) \href{https://doi.org/10.1093/gigascience/gix101}{Gigascience 6:12}
}
\keyword{datasets}
