% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{multiMerge}
\alias{multiMerge}
\title{Merge multiple data.frames}
\usage{
multiMerge(list, columns = list())
}
\arguments{
\item{list}{vector of data.frame names}

\item{columns}{list}
}
\value{
data.frame
}
\description{
Merge the given list of data frame into one dataframe
}

