% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{query_column_keywords}
\alias{query_column_keywords}
\title{query_column_keywords}
\usage{
query_column_keywords(dt, table_name, keyword_string)
}
\arguments{
\item{dt}{data.frame}

\item{table_name}{character}

\item{keyword_string}{character}
}
\description{
return a vector of column names, given a table name and keyword string
}
\examples{
\dontrun{
dt <- lagos_load("1.087.1")
query_column_keywords(dt, "hu12.chag", "hydrology")
}
}
