% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{query_column_names}
\alias{query_column_names}
\title{query_column_names}
\usage{
query_column_names(dt, table_name, grep_string)
}
\arguments{
\item{dt}{data.frame}

\item{table_name}{character}

\item{grep_string}{character}
}
\description{
return a vector of column names, given a table name and grep
 query string
}
\examples{
\dontrun{
dt <- lagos_load("1.054.1")
query_column_names(dt, "hu4.chag", "_dep_")
}
}

