% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{query_lagos_names}
\alias{query_lagos_names}
\title{query_lagos_names}
\usage{
query_lagos_names(dt, grep_string)
}
\arguments{
\item{dt}{data.frame output of \code{\link[LAGOS]{lagos_load}}}

\item{grep_string}{character search string to grep to table column names}
}
\description{
return table names with column names that grep to query
}
\examples{
query_lagos_names(lagos_load("1.054.1"), "dep")
}

