% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{readCol}
\alias{readCol}
\title{Return the position of a column name?}
\usage{
readCol(name, geo, limno, info)
}
\arguments{
\item{name}{character?}

\item{geo}{data.frame output of \code{\link[LAGOS]{lagos_compile}}}

\item{limno}{data.frame output of \code{\link[LAGOS]{lagos_compile}}}

\item{info}{data.frame output of \code{\link[LAGOS]{lagos_compile}}}
}
\description{
Return the position of a column name?
}

