% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOS-package.R
\docType{data}
\name{secchi}
\alias{secchi}
\title{Secchi (Water Clarity) Data}
\format{A data frame with 913319 observations of 12 variables:
\itemize{
     \item eventida1087: unique combination of programid, lakeid, and date
     for each sampling event
     \item lagoslakeid: unique LAGOS lake identifier
     \item programname: name of the sampling/monitoring program that
     collected the data
     \item programtype: categorical description of the type of sampling
     /monitoring effort
     \item lagosversion: current version of LAGOS that the data record
     belongs to
     \item sampledate: date at which the sample was collected, stored in date format (YYYY-MM-DD)
     \item secchi: Secchi depth in meters
     \item secchi_censorcode: identifies whether a Secchi value is censored
     and the censor type
     \item secchi_qual: data flags from the source program
     \item secchi_methodinfo: flag to indicate variables that were sampled
     using different methods. "SECCHI_VIEW" =
     viewing box used to measure Secchi depth, "SECCHI_VIEW_UNKNOWN" =
     unknown whether viewing box was used, "NA" = no viewing
     box used to measure Secchi depth.
     \item sampleyear: year (4-digit) of sampling event
     \item samplemonth: month (1 through 12) of sampling event
     }}
\description{
A dataset containing Secchi disk measurements from lakes.
}
\details{
There is some overlap between the Secchi values reported in table 'secchi'
and table 'epi_nutr'. The 'epi_nutr' table only contains Secchi observations
when there were additional nutrient data collected at the same time.
}
\references{
Soranno et al. (2015) \href{https://gigascience.biomedcentral.com/articles/10.1186/s13742-015-0067-4}{Gigascience 4:28}
}
\keyword{datasets}
