% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{sortList}
\alias{sortList}
\title{Sort a list}
\usage{
sortList(list, info)
}
\arguments{
\item{list}{vector of data.frame names}
}
\value{
list
}
\description{
Sort the given list in a way that merging computation cost less
}

