% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOS-package.R
\docType{data}
\name{state}
\alias{state}
\title{State Data}
\format{A data frame with 17 observations of 7 variables:
\itemize{
    \item state: abbreviation of state name
    \item state_name: text name of state
    \item state_zoneid: unique zone identifier
    \item state_lat: latitude of centroid of each polygon in decimal degrees (NAD83)
    \item state_long: longitude of centroid of each polygon in decimal degrees (NAD83)
    \item state_pct_in_nwi: percent of polygon within the USFWS National Wetland Inventory (NWI) boundary
    \item state_ha_in_nwi: area (ha) of polygon within the USFWS National Wetland Inventory (NWI) boundary
    }}
\description{
A dataset containing information about state characteristics for LAGOS lakes.
For details on zone delineations, see additional file 7 in Soranno et al. (2015).
}
\references{
Soranno et al. (2015) \href{https://gigascience.biomedcentral.com/articles/10.1186/s13742-015-0067-4}{Gigascience 4:28}
}
\keyword{datasets}

