% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{stepsfunc}
\alias{stepsfunc}
\title{stepsfunc}
\usage{
stepsfunc(list, columns, identifiers)
}
\arguments{
\item{list}{list?}

\item{columns}{numeric?}

\item{identifiers}{character?}
}
\description{
Loops through a list and does something...?
}

