% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{info_table}
\alias{info_table}
\title{Summarize all LAGOSNE flat files}
\usage{
info_table(lg)
}
\arguments{
\item{lg}{list output of lagosne_load}
}
\description{
Generate table summary statistics for all LAGOSNE data frames.
}
\examples{
\dontrun{
lg <- lagosne_load("1.087.3")
LAGOSNE:::info_table(lg)
}
}
\author{
Masrour Farzan
}
