% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{info_table}
\alias{info_table}
\title{info_table}
\usage{
info_table(geo, limno)
}
\arguments{
\item{geo}{data.frame}

\item{limno}{data.frame}
}
\description{
generate table summary statistics
}
\author{
Masrour Farzan
}

