% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOS-package.R
\docType{data}
\name{iws}
\alias{iws}
\title{Watershed Data}
\format{A data frame with 51065 observations and 12 variables:
\itemize{
    \item iws_nhdid: permanent lake-specific identifier from the National Hydrography Dataset
    \item iws_areaha: area (ha) of IWS
    \item iws_perimkm: perimeter (km) of IWS
    \item iws_lat: latitude of centroid of IWS polygon in decimal degrees (NAD83)
    \item iws_long: longitude of centroid of IWS polygon in decimal degrees (NAD83)
    \item iws_lakeareaha: lake area (ha) of focal lake within iws polygon
    \item iws_pct_in_usa: percent of polygon within the US boundary
    \item iws_ha_in_usa: area (ha) of polygon within the US boundary
    \item iws_pct_in_nwi: percent of polygon within the USFWS National Wetland Inventory (NWI) boundary
    \item iws_ha_in_nwi: area (ha) of polygon within the USFWS National Wetland Inventory (NWI) boundary
    \item iws_zoneid: a unique identifier for each IWS in LAGOS
    \item iws_lagoslakeid: a unique identifier for each lake in LAGOS
    }}
\description{
A dataset containing information on the interlake watershed (IWS) characteristics for LAGOS lakes.
For details on how the watersheds were delineated, see additional file 8 in Soranno et al. (2015).
}
\references{
Soranno et al. (2015) \href{https://gigascience.biomedcentral.com/articles/10.1186/s13742-015-0067-4}{Gigascience 4:28}
}
\keyword{datasets}

