% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{lagos_compile}
\alias{lagos_compile}
\title{Compile LAGOS data}
\usage{
lagos_compile(version)
}
\arguments{
\item{version}{character LAGOS database version string}
}
\description{
Compile LAGOS data from component flat files
}
\examples{
\dontrun{
lagos_compile("1.054.1")
}

}

