% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{lagos_load}
\alias{lagos_load}
\title{lagos_load}
\usage{
lagos_load(version, format = "rds", fpath = NA)
}
\arguments{
\item{version}{character LAGOS database version string}

\item{format}{character choice of rds or sqlite}

\item{fpath}{file.path optionally specify custom location of LAGOS rds file}
}
\description{
Load LAGOS files from local file system
}
\examples{
\dontrun{
dt  <- lagos_load("1.054.1")
dt2 <- lagos_load("1.054.2")

compare_columns <- function(num){
 identical(dt2$epi.nutr[,num], dt$epi.nutr[,num])
}
names(dt2$epi.nutr)[sapply(13:94, compare_columns)]

}
}

