% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{lagos_load}
\alias{lagos_load}
\title{lagos_load}
\usage{
lagos_load(version, fpath = NA)
}
\arguments{
\item{version}{character LAGOS database version string}

\item{fpath}{file.path optionally specify custom location of LAGOS rds file}
}
\description{
Load LAGOS files from local file system
}
\examples{
\dontrun{
dt <- lagos_load("1.054.1")
}
}

