% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{lagos_load_oliver_2015}
\alias{lagos_load_oliver_2015}
\title{lagos_load_oliver_2015}
\usage{
lagos_load_oliver_2015(fpath = NA)
}
\arguments{
\item{fpath}{file.path optionally specify custom location of csv data file}
}
\description{
lagos_load_oliver_2015
}
\examples{
\dontrun{
lagos_load_oliver_2015()
}
}
