% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{lagos_select}
\alias{lagos_select}
\title{lagos_select}
\usage{
lagos_select(dt, scale = NULL, category = NULL,
  table_column_nested = NULL)
}
\arguments{
\item{dt}{data.frame}

\item{scale}{character filter results by one of:
\itemize{
    \item county
    \item edu
    \item hu4
    \item hu8
    \item hu12
    \item state
}}

\item{category}{character}

\item{table_column_nested}{nested list of tables and table-columns to select from LAGOS}
}
\description{
Select and merge columns based on a nested list of LAGOS tables and column names.
}
\examples{
\dontrun{
dt <- lagos_load("1.054.1")

# select only from limno
dt_reduced <- lagos_select(dt, table_column_nested =
                list("epi.nutr" = c("tp", "tn")))

# select only from geo
dt_reduced <- lagos_select(dt, table_column_nested =
                list("iws.lulc" = c("iws_nlcd2011_pct_95")))

# select individual table-columns from limno and geo
table_columns <- list("epi.nutr" = c("tp", "tn"),
                   "iws.lulc" = c("iws_nlcd2011_pct_95"))
dt_reduced    <- lagos_select(dt, table_column_nested = table_columns)

# select from specific tables using keywords
dt_reduced <- lagos_select(dt, table_column_nested = list("epi.nutr" = "waterquality"))

# select from multiple specific tables using keywords
table_columns <- list("epi.nutr" = c("waterquality"),
                   "hu4.chag" = c("deposition"))
dt_reduced    <- lagos_select(dt, table_column_nested = table_columns)

# select based non-specific tables using keywords
dt_reduced <- lagos_select(dt, scale = "HU4",
               category = c("waterquality", "deposition"))

# select based on inexact keywords and exact table specification
table_columns <- list("epi.nutr" = c("doc", "lagoslakeid"))
dt_reduced    <- lagos_select(dt, scale = "HU4", category = c("deposition"),
                  table_column_nested = table_columns)
}
}

