% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{lagos_select}
\alias{lagos_select}
\title{lagos_select}
\usage{
lagos_select(dt, table_column_nested)
}
\arguments{
\item{dt}{data.frame}

\item{table_column_nested}{nested list of tables and table-columns to select from LAGOS}
}
\description{
Select and merge columns based on a nested list of LAGOS tables and column names.
}
\examples{
\dontrun{
dt <- lagos_load("1.054.1")

# select individual table-columns
table_columns <- list("epi.nutr" = c("tp", "tn"), "iws.lulc" = c("iws_nlcd2011_pct_95"))
dt_reduced <- LAGOS::lagos_select(dt, table_columns)

# group-select using keywords
table_columns <- list("epi.nutr" = c("waterquality", "lagoslakeid"), "hu4.chag" = c("deposition"))
dt_reduced <- LAGOS::lagos_select(dt, table_columns)
}
}

