% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSNE-package.R
\docType{package}
\name{LAGOSNE-package}
\alias{LAGOSNE-package}
\alias{LAGOSNE}
\title{R interface to the LAGOSNE lakes database}
\description{
Interface to the LAGOSNE lakes database
}
\author{
\email{stachel2@msu.edu}
}
