% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{lagosne_get}
\alias{lagosne_get}
\title{Retrieve LAGOSNE flat files}
\usage{
lagosne_get(version = lagosne_version(), overwrite = FALSE,
  dest_folder = tempdir())
}
\arguments{
\item{version}{character LAGOSNE database version string}

\item{overwrite}{logical overwrite existing data for the specified version}

\item{dest_folder}{file.path to save data. Default to a temporary folder.
Recommended to set to LAGOSNE:::lagos_path() so that data persists between
R sessions.}
}
\description{
Retrieve LAGOSNE flat files from EDI.
}
\examples{
\dontrun{
lagosne_get()
}
}
