% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{lake_info}
\alias{lake_info}
\title{Get a summary of lake attribute information}
\usage{
lake_info(dt, name = NA, state = NA, lagoslakeid = NA)
}
\arguments{
\item{dt}{list of data.frames. output of \code{\link[LAGOSNE]{lagosne_load}}.}

\item{name}{character lake name not caps sensitive}

\item{state}{character state name not caps sensitive}

\item{lagoslakeid}{numeric lake id}
}
\description{
Return lake attribute information from its name and state or its lagoslakeid.
}
\examples{
\dontrun{
dt <- lagosne_load("1.087.1")
lake_info(dt, lagoslakeid = 4314)
lake_info(dt, lagoslakeid = 7441)
lake_info(dt, lagoslakeid = 244)
lake_info(dt, lagoslakeid = 4686)
lake_info(dt, lagoslakeid = 8016)
lake_info(dt, lagoslakeid = c(1441))
lake_info(dt, lagoslakeid = c(125428, 1441))
lake_info(dt, lagoslakeid = c(4686, 8016))

lake_info(dt, "Sunapee Lake", "New Hampshire")
lake_info(dt, name = c("Sunapee Lake", "Oneida Lake"),
              state = c("New Hampshire", "New York"))
}
}
