% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{lake_info}
\alias{lake_info}
\title{lake_info}
\usage{
lake_info(dt, name, state)
}
\arguments{
\item{dt}{data.frame output of \code{\link[LAGOS]{lagos_load}}}

\item{name}{character lake name not caps sensitive}

\item{state}{character state name not caps sensitive}
}
\description{
lake_info
}
\examples{
\dontrun{
dt <- lagos_load("1.054.1")

  name = c("Oneida Lake", "Sunapee Lake", "Lake Mendota"),
  state = c("New York", "New Hampshire", "Wisconsin"))

  apply(focal_lakes, 1, function(x) lake_info(
    dt = dt$limno$lake.specific, name = x[1], state = x[2]))
}
}

