% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSNE-package.R
\docType{data}
\name{lg_extent}
\alias{lg_extent}
\title{LAGOSNE extent}
\format{An object of class \code{sfc_POLYGON} (inherits from \code{sfc}) of length 1.}
\usage{
lg_extent
}
\description{
A simple features polygons representing the geographical extent of the LAGOSNE dataset
}
\references{
Soranno et al. (2017) \href{https://doi.org/10.1093/gigascience/gix101}{Gigascience 6:12}
}
\keyword{datasets}
