% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSNE-package.R
\docType{data}
\name{lg_extent}
\alias{lg_extent}
\title{The spatial extent of LAGOSNE data}
\format{
A single sf polygon
}
\usage{
lg_extent
}
\description{
A dataset containing the spatial extent of LAGOSNE data.
}
\keyword{datasets}
