% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOS-package.R
\docType{data}
\name{lg_subset}
\alias{lg_subset}
\title{LAGOS subset}
\format{An object of class \code{list} of length 26.}
\usage{
lg_subset
}
\description{
A dataset with the first two rows of many LAGOS tables
}
\references{
Soranno et al. (2015) \href{https://gigascience.biomedcentral.com/articles/10.1186/s13742-015-0067-4}{Gigascience 4:28}
}
\keyword{datasets}
