% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_lagos_txt}
\alias{load_lagos_txt}
\title{load_lagos_txt}
\usage{
load_lagos_txt(file_name, sep = "\\t", ...)
}
\arguments{
\item{file_name}{character}

\item{sep}{character seperator (tab or comma seperated values)}

\item{...}{Options passed on to \code{\link[utils]{read.table}}}
}
\value{
data.frame
}
\description{
A wrapper for \code{\link[utils]{read.table}} with a default set of parameters.
}
