% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{multiMerge}
\alias{multiMerge}
\title{Merge multiple data.frames}
\usage{
multiMerge(informationTable, list, columns = list())
}
\arguments{
\item{informationTable}{data.frame info table from \code{\link[LAGOS]{lagos_compile}}}

\item{list}{vector of data.frame names}

\item{columns}{list}
}
\value{
data.frame
}
\description{
Merge the given list of data frame into one dataframe
}
\examples{
\dontrun{
dt <- lagos_load("1.054.1")
tables_list <- c("iws.lulc", "epi.nutr", "iws.conn", "hu4.chag")
}

}

