% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{query_column_keywords}
\alias{query_column_keywords}
\title{query_column_keywords}
\usage{
query_column_keywords(dt, table_name, keyword_string)
}
\description{
return a vector of column names, given a table name and keyword string
}
\examples{
\dontrun{
dt <- lagos_load("1.054.1")
query_column_keywords(dt, "epi.nutr", "waterquality")
}
}

