% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{query_column_names}
\alias{query_column_names}
\title{Query columnnames}
\usage{
query_column_names(dt, table_name, grep_string)
}
\arguments{
\item{dt}{data.frame}

\item{table_name}{character}

\item{grep_string}{character}
}
\description{
Return a vector of column names, given a table name and grep query string.
}
\examples{
\dontrun{
dt <- lagosne_load("1.087.3")
query_column_names(dt, "hu4.chag", "_dep_")
query_column_names(dt, "county.chag", "baseflowindex")
}
}
