% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{query_lagos_names}
\alias{query_lagos_names}
\title{Query LAGOSNE names}
\usage{
query_lagos_names(grep_string, scale = NA, dt = lagosne_load())
}
\arguments{
\item{grep_string}{character search string to grep to table column names}

\item{scale}{character filter results by one of:
\itemize{
    \item county
    \item edu
    \item hu4
    \item hu8
    \item hu12
    \item state
}}

\item{dt}{data.frame output of \code{\link[LAGOSNE]{lagosne_load}}}
}
\description{
Return a vector of table names whose associated tables have
 columns that grep to query.
}
\examples{
\dontrun{
query_lagos_names("_dep_")
query_lagos_names("_dep_", "hu4")
query_lagos_names("chla")
query_lagos_names("secchi")
query_lagos_names("conn")
}
}
