% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSNE-package.R
\docType{data}
\name{source}
\alias{source}
\alias{programs}
\alias{lagos.source.programs}
\alias{source.programs}
\alias{lagos.source}
\title{LAGOSNE sources}
\format{A data frame with 87 observations of 19 variables:
\itemize{
     \item programid: unique LAGOSNE program identifier
     \item programname: name of the sampling/monitoring program that
     collected the data
     \item metadataid: unique metadata identifier associated with
     \item sourceid
     \item sourcename
     \item composite
     \item programtype: categorical description of the type of sampling
     /monitoring effort
     \item fundingsource
     \item labtype
     \item programlink
     \item programstatus
     \item dataprovider
     \item organization
     \item sourcedescription
     \item beg_yr
     \item end_yr
     \item state_name
     \item lakecount
     \item variable_list
     }}
\description{
A dataset identifying the sources of limnological data in LAGOSNE.
}
\references{
Soranno et al. (2015) \href{https://gigascience.biomedcentral.com/articles/10.1186/s13742-015-0067-4}{Gigascience 4:28}
}
\keyword{datasets}
